Add baseline defaults section [ci skip]
Mentioning the baseline default for a config option can be confusing
when that default is overridden by `config.load_defaults`.  To avoid
that confusion, this commit relocates such baseline defaults from their
explanatory paragraphs to a "Baseline defaults" section that flows with
the other `config.load_defaults` sections.

Ideally, when we override other baseline defaults in the future, we will
relocate mention of them as done here.

Closes #39387.