Add db_runtime to Active Job instrumentation
This adds `db_runtime` to the notification payload of a
`"perform.active_job"` event.  `db_runtime` tracks the total time taken
by database queries while performing a job, which helps in understanding
how a job's time is spent.  This is similar to what is done for
controller actions in `ActiveRecord::Railties::ControllerRuntime`.

Closes #35354.

Co-authored-by: Cory Gwin <gwincr11@github.com>