update generation instructions for consistency
New to Rails but looks like everything for an engine should be done inside the engine's folder. The updated text has been there for a long time so I hope I'm not just missing something. At this point in the example the user does not have an application (that comes in the next section) and elsewhere in the doc there is talk of the importance of running the commands from the *engine* root so it seems more consistent to talk about "engine root" as updated.