Rename horrible and dirty to puzzled and confused in test models
In 7f938ca the test model `Man` was renamed to `Human`. Maybe this is a
good time to also change `horrible_human` and `dirty_human` to
`happy_human` and `confused_human`.

While this change is mostly cosmetic change, the phrase "dirty man" has
a negative meaning.

The adjectives "confused" and "puzzled" were chosen because they are
used for defining associations with errors.