Allow false as check_box unchecked value
When using boolean false as value for `check_box` unchecked value, the
hidden input for unchecked value is not generated.

Presently, the following example:

```ruby
<%= form_with :model do |form_builder| %>
  <%= form_builder.check_box :attribute, {}, true, false %>
<% end %>
```

will only output the input for true, ie:

```html
<input id="model_attribute name="model[attribute]" type="checkbox" value="true" />
```

instead of the expected behaviour:

```html
<input name="model[attribute]" type="hidden" value="false" />
<input id="model_attribute name="model[attribute]" type="checkbox" value="true" />
```

This is somewhat surprising behaviour, in particular since a different
order of values will work correctly, ie:

```ruby
<%= form_with :model do |form_builder| %>
  <%= form_builder.check_box :attribute, {}, false, true %>
<% end %>
```

as well as using strings:

```ruby
<%= form_with :model do |form_builder| %>
  <%= form_builder.check_box :attribute, {}, "false", "true" %>
<% end %>
```

This is due to a ternary operator being used to determine if an
unchecked value is present. On receiving a boolean false, an empty
string is returned.

This PR changes the behaviour to cast the unchecked value to a string
and calling `.present?` on it, so as to allow the expected behaviour to
occur.