Fix TimeWithZone to work with Time.at
To use the precise Rational representation of time, Time.at requires
`to_int` to be defined in addition to `to_r` to make a distinction and
avoid using `to_r` with objects that are not time objects, e.g. a
String "123.456" (Rational(15432/125)).

https://github.com/ruby/ruby/commit/7e1fddba4a609cb7bf4a696eccd892e68753bb21
https://github.com/ruby/ruby/blob/7e1fddba4a609cb7bf4a696eccd892e68753bb21/spec/ruby/core/time/at_spec.rb#L96
https://bugs.ruby-lang.org/issues/17131#note-1

Time.at performs a check if the first argument is kind of a `::Time`,
however it can't be tricked by overriding `kind_of?`, and it wouldn't be
sufficient, since it uses the internal structure of the `Time` class to
copy over the time data.

In case when `to_r` is defined and `to_int` is not, it fails fast with a
TypeError. This was previously considered in a test (initially
introduced by
https://github.com/rails/rails/commit/b7f9de27f0558d6144f982cae83f32ca85a07f7e.
The nature of origin of the exception was not determined at the moment.

This commit extends the change made in #9403