Make Time.at loseless with TimeWithZone argument
Float representation is not sufficient to preserve nanosecond precision.
Previously, `to_f` conversion was needed because TimeWithZone failed to
conform to `Time.at` protocol that requires the object to respond to
`to_int` in addition to `to_r`.
https://bugs.ruby-lang.org/issues/17131#note-1

For `DateTime`, the coercion to `Time` is expensive, but most probably
comparable to the effort needed to coerce it to floating point number
and back to `Time` that `Time.at` does internally.

fixes #38831