Improve ActiveRecord::Base.destroy_later implementation
* Fix CHANGELOG entry, improve documentation

Additionally, add docs about how to configure destroy jobs

* Avoid using abbreviations in argument names

* Check if job is configured when calling destroy_later on class

* Pass keyword arguments to the job to avoid warnings on Ruby 2.7

* Move active record queues configuration to Active Record

This config is related to jobs but it doesn't load anything on Active
Job, so there is no need to move the config to the Active Job railtie.

* Allow configuring the destroy later jobs per model

We are still going to configure this globally to the framework but if
a user wants to use a different job in a model the system will work.
This also make easier to isolate the Active Job tests from the Active
Record jobs.

* Change the tests description to talk about the framework features

The tests were talking about the model features, not which features in
the framework they were asserting.

Co-authored-by: Rafael Mendonça França <rafael@franca.dev>
Co-authored-by: Cory Gwin @gwincr11 <gwincr11@github.com>