Allow extra client config options for ActiveStorage::Service::AzureStorage
This is a backport of updates made to [master](https://github.com/rails/rails/pull/35096).

In order to set custom options (e.g. a custom domain name) on the azure storage client, the ActiveStorage adapter must allow more options to be passed to the initializer. Currently `ActiveStorage::Service::AzureStorage` only allows for a fixed set of options.

Here we add the ability to pass additional options to the initializer.