Fix application templates on existing projects
When running an application template on a new project, we get the
following:
1. template is applied and gems are added to the Gemfile
   https://guides.rubyonrails.org/rails_application_templates.html#gem-args
2. `bundle install` is run
3. The block passed to `after_bundle` is executed allowing you to run
commands added by gems included in the template.

The documentation for application templates states that they can be run
on existing projects via the syntax
(https://guides.rubyonrails.org/rails_application_templates.html#usage):

`rails app:template LOCATION=~/template.rb`

However if you were to try this neither `bundle install` or the
`after_bundle` callback will be executed which makes this far less
useful.

This change adds the required calls so that application templates are
consistent even when added to an existing project.