Backport: Disentangle Action Text from ApplicationController
The original commit was merged to the 6.1-stable branch.
It also fixes #36810 on the 6.0-stable branch.

Original commit message:
This commit allows Action Text to be used without having an
ApplicationController defined.  In doing so, it also fixes Action Text
attachments to render the correct URL host in mailers.

It also avoids allocating an ActionController::Renderer per request.

Co-authored-by: Jeremy Daer <jeremydaer@gmail.com>