Ensure `connects_to` can only be called on base or abstract classes
`connectes_to` should only be called on `ActiveRecord::Base` or abstract
classes. This is recommended in the documentation but until now was not
enforced by the code. It's unsafe to open too many connections to mysql
(and probably other databases), so it's safest to have 1 class for the
connection and subclass from that.

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>