Serialize PostgreSQL composite types into Ruby hashes
Since 2014 there has been no special support for PostgreSQL composite types.
They were mapped to normal text columns, leaving parsing of data
at the discretion of the Rails application.

In this commit this behavior is changed so that values are mapped
into hashes indexed by names of the corresponding attributes.
For backward compatibility this setting is turned off. To turn it on,
you should set `config.active_record.pg_composite_type_cast = :hash`:

    # db/migrate/20201017105000_create_contacts.rb
    execute <<-SQL
      CREATE TYPE full_address AS
       (
         city VARCHAR(90),
         street VARCHAR(90)
       );
    SQL
    create_table :contacts do |t|
      t.column :address, :full_address
    end

    # config/application.rb
    class Application < Rails::Application
      config.active_record.pg_composite_type_cast = :hash
    end

    # app/models/contact.rb
    class Contact < ApplicationRecord
    end

    # Usage
    Contact.create address: { city: "Paris", street: "Champs-Élysées" }
    contact = Contact.first
    contact.address # => { city: "Paris", street: "Champs-Élysées" }

The resulting hashes can be nested when some attributes are of
composite types by themselves.

To provide smoother migration to the new behavior, you can use
plain old strings to assign value to the attribute.

    contact.address = "(London,Trafalgar Square)"
    contact.address # => { city: "London", street: "Trafalgar Square" }

Any object responding to the `to_h` method call can be used
as an attribute value as well:

    contact.address = Struct.new(:city, :street)["Amsterdam", "Damrak"]
    contact.address # => { city: "Amsterdam", street: "Damrak" }

When a new setting `pg_composite_type_cast` wasn't changed to `:hash`,
the behavior of the serializer is kept unchanged:

    contact.address # => "(Amsterdam,Damrak)"