Use non-memoized route helper modules at namespace boundaries.
A possible fix for the failing test in #40263, hopefully without
throwing out the gains of #37927.

If we're switching namespace in a controller subclass (especially
due to an isolated engine controller appearing in the ancestors)
then we may need a fresh route-helper module to "mask" the namespaced
routes.

The idea here is, at the point where a module would be included,
ask the ancestors what their namespace was, and if ours is different
then assume a fresh module is required.

If not, then do not ask for a fresh module, to avoid trampling on
the intentions of #37927.  It looks like validating benchmarks were
run at the time, it might be helpful to run them against this change
to avoid performance regression.

This approach relies on the ancestors for any given controller class
being essentially invariant (which isn't perfectly guaranteed but
may be true enough); and, permission to add an additional parameter
to ActionDispatch::Routing::RouteSet#url_helpers, which needs review
from someone more familiar than me with the routing engine.