Add a note about PR approvals [ci skip]
Sometimes folks get "Approved" reviews from other contributors who do
not have merge access. This can come across as an official endorsement,
especially in cases where the approval comes from someone who is a
member of the Rails org, but is not on the core or committer team (me,
for example).

This paragraph is meant to document both that "Approved" is not the same
as an official endorsement, and that we prefer people don't mark PRs as
"Approved" in the first place.