Only set backreferences in gsub/sub if argument is a Regexp
Setting backreferences in the block passed to sub/gsub on
ActiveSupport::SafeBuffer incurs a significant performance penalty. If
the argument to sub/gsub is a string, there are no backreferences so
this penalty is paid for nothing.

We can avoid this penalty in this case simply by checking whether the
first argument passed to the method is a Regexp.