Fix initializing STI from abstract class
It should be possible to call `.new` on the abstract class and return
the subclass instantiated properly just like it would with a base class.

Given the following example:

```
class AbstractCompany < ActiveRecord::Base
  self.abstract_class = true
end

class Company < AbstractCompany
end

class Acme < Company
end
```

It is possible to calls `Company.new(type: "Acme")` and it will return
an instance of Acme.

However, if someone would call `AbstractCompany.new(type: "Acme")`, it
would raise a `NotImplementedError`.

This fixes it so that calling new on the abstract class behaves the
same.